Move paymentDTO to order module
