+++ src/core/orders/interfaces/UpdateOrderPaymentStatus.d.ts
@@ -1,4 +1,4 @@
-import PaymentDTO from "../../webhooks/dto/PaymentDTO";
+import PaymentDTO from "../dto/PaymentDTO";
 
 export default interface UpdateOrderPaymentStatus {
   updateOrderPaymentStatus(paymentDTO: PaymentDTO): Promise<OrderDTO>;
